Hardcode specific version of deps for example app
Note a new version of asp-net-core has been released v5.0.0. As of this
release, asp-net-core is a part of dot-net, so to get asp-net-core, you
need to install dot-net v5.0.0.

The dependency which we provide to link asp-net-core to MySQL/MariaDB
does not yet have a stable release compatible with v5.

https://www.nuget.org/packages/Pomelo.EntityFrameworkCore.MySql/